filter addAdmins tab by current used admins
