+++ src/pages/domain/DomainAdminsPage.tsx
@@ -28,8 +28,7 @@ import CONST from '@src/CONST';
 import ONYXKEYS from '@src/ONYXKEYS';
 import ROUTES from '@src/ROUTES';
 import type SCREENS from '@src/SCREENS';
-import type * as OnyxTypes from '@src/types/onyx';
-import getEmptyArray from '@src/types/utils/getEmptyArray';
+import selectAdminIDs from '@src/libs/DomainUtils';
 
 type DomainAdminsPageProps = PlatformStackScreenProps<DomainSplitNavigatorParamList, typeof SCREENS.DOMAIN.SAML>;
 
@@ -37,23 +36,6 @@ type AdminOption = Omit<ListItem, 'accountID' | 'login'> & {
     accountID: number;
     login: string;
 };
-
-// to be moved to utils
-function selectAdminIDs(domain: OnyxTypes.Domain | undefined): number[] {
-    if (!domain) {
-        return [];
-    }
-
-    return (
-        Object.entries(domain)
-            .filter(([key]) => key.startsWith(ONYXKEYS.COLLECTION.DOMAIN_ADMIN_PERMISSIONS))
-            .map(([, value]) => {
-                const rawValue = typeof value === 'object' && value !== null && 'value' in value ? value.value : value;
-                return Number(rawValue);
-            })
-            .filter((id) => !Number.isNaN(id)) ?? getEmptyArray<string>()
-    );
-}
 
 function DomainAdminsPage({route}: DomainAdminsPageProps) {
     const domainID = route.params.accountID;
